Move class to the Pages module
