+++ src/Backend/Modules/Extensions/Installer/Installer.php
@@ -3,7 +3,7 @@
 namespace Backend\Modules\Extensions\Installer;
 
 use Backend\Core\Installer\ModuleInstaller;
-use Common\ModuleExtraType;
+use Backend\Modules\Pages\Domain\ModuleExtra\ModuleExtraType;
 
 /**
  * Installer for the extensions module.
