+++ src/Backend/Modules/FormBuilder/Engine/Model.php
@@ -5,7 +5,7 @@ namespace Backend\Modules\FormBuilder\Engine;
 use Backend\Core\Engine\Model as BackendModel;
 use Backend\Core\Language\Language as BL;
 use Backend\Modules\Pages\Domain\ModuleExtra\ModuleExtraRepository;
-use Common\ModuleExtraType;
+use Backend\Modules\Pages\Domain\ModuleExtra\ModuleExtraType;
 use Frontend\Core\Language\Language as FL;
 use Symfony\Component\Finder\Finder;
