+++ src/Backend/Modules/Search/Installer/Installer.php
@@ -4,9 +4,9 @@ namespace Backend\Modules\Search\Installer;
 
 use Backend\Core\Engine\Model as BackendModel;
 use Backend\Core\Installer\ModuleInstaller;
+use Backend\Modules\Pages\Domain\ModuleExtra\ModuleExtraType;
 use Backend\Modules\Pages\Domain\PageBlock\PageBlock;
 use Backend\Modules\Pages\Domain\PageBlock\PageBlockRepository;
-use Common\ModuleExtraType;
 
 /**
  * Installer for the search module
